pyproject.toml
[project]
name = "conference-map"
version = "0.1.0"
description = "Conference topic visualization using BERTopic and datamapplot"
requires-python = ">=3.9"
dependencies = [
    "bertopic>=0.16.0",
    "sentence-transformers>=2.2.0",
    "umap-learn>=0.5.0",
    "datamapplot>=0.2.0",
    "scikit-learn>=1.0.0",
    "pillow>=9.0.0",
    "plotly>=5.0.0",
    "kaleido>=0.2.0",
    "matplotlib>=3.5.0",
    "ipywidgets>=8.0.0",
    "anywidget>=0.9.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
